Packages refactor
