+++ src/main/java/ru/mail/park/gwent/Application.java
@@ -7,7 +7,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.socket.WebSocketHandler;
 import org.springframework.web.socket.handler.PerConnectionWebSocketHandler;
-import ru.mail.park.gwent.websocket.GameWebSocketHandler;
+import ru.mail.park.gwent.websocket.handler.GameWebSocketHandler;
 
 @SpringBootApplication
 public class Application {
