+++ src/main/java/ru/mail/park/gwent/websocket/message/ClientState.java
@@ -1,4 +1,4 @@
-package ru.mail.park.gwent.websocket;
+package ru.mail.park.gwent.websocket.message;
 
 public enum ClientState {
     INIT(0),
