+++ src/main/java/ru/mail/park/gwent/websocket/message/WantPlayMessage.java
@@ -1,4 +1,4 @@
-package ru.mail.park.gwent.websocket;
+package ru.mail.park.gwent.websocket.message;
 
 public class WantPlayMessage extends WebSocketMessage {
     private ClientState state;
