+++ src/test/java/ru/mail/park/gwent/controllers/SessionControllerTest.java
@@ -9,9 +9,9 @@ import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.*;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.context.junit4.SpringRunner;
-import ru.mail.park.gwent.domains.Message;
-import ru.mail.park.gwent.domains.UserInfo;
-import ru.mail.park.gwent.domains.UserProfile;
+import ru.mail.park.gwent.domains.auth.Message;
+import ru.mail.park.gwent.domains.auth.UserInfo;
+import ru.mail.park.gwent.domains.auth.UserProfile;
 import ru.mail.park.gwent.services.UserService;
 
 import java.util.Collections;
@@ -26,7 +26,7 @@ import static org.mockito.Mockito.when;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 import static ru.mail.park.gwent.consts.Constants.AUTH_URL;
 import static ru.mail.park.gwent.consts.Constants.LOCALHOST;
-import static ru.mail.park.gwent.domains.MessageEnum.*;
+import static ru.mail.park.gwent.domains.auth.MessageEnum.*;
 
 /**
  * @author Konstantin Gulyy
