+++ src/test/java/ru/mail/park/gwent/controllers/UserControllerTest.java
@@ -11,8 +11,8 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.junit4.SpringRunner;
-import ru.mail.park.gwent.domains.Message;
-import ru.mail.park.gwent.domains.UserProfile;
+import ru.mail.park.gwent.domains.auth.Message;
+import ru.mail.park.gwent.domains.auth.UserProfile;
 import ru.mail.park.gwent.services.UserService;
 
 import java.util.Collections;
@@ -23,7 +23,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 import static ru.mail.park.gwent.consts.Constants.SIGN_UP_URL;
-import static ru.mail.park.gwent.domains.MessageEnum.*;
+import static ru.mail.park.gwent.domains.auth.MessageEnum.*;
 
 /**
  * @author Konstantin Gulyy
